+++ src/Plugin/AmpacheGoogleAnalytics.php
@@ -63,8 +63,7 @@ class AmpacheGoogleAnalytics
             return false;
         }
 
-        Preference::insert('googleanalytics_tracking_id', T_('Google Analytics Tracking ID'), '', 100, 'string',
-            'plugins', $this->name);
+        Preference::insert('googleanalytics_tracking_id', T_('Google Analytics Tracking ID'), '', 100, 'string', 'plugins', $this->name);
 
         return true;
     }